Unify project structure for the Arcane service
