+++ src/Configurations/StreamClassOperatorServiceConfiguration.cs
@@ -1,5 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
-using Arcane.Operator.Services.Operator;
+using Arcane.Operator.Services.Operators;
 
 namespace Arcane.Operator.Configurations;
