+++ src/Models/Commands/StreamClassCommands.cs
@@ -1,8 +1,10 @@
 ﻿using Arcane.Operator.Models.Api;
+using Arcane.Operator.Models.Base;
 using Arcane.Operator.Models.Resources;
 using Arcane.Operator.Models.Resources.Status.V1Alpha1;
 using Arcane.Operator.Models.Resources.StreamClass.Base;
 using Arcane.Operator.Services.Base;
+using Arcane.Operator.Services.Base.CommandHandlers;
 
 namespace Arcane.Operator.Models.Commands;
