+++ src/Models/Commands/StreamDefinitionCommands.cs
@@ -1,6 +1,8 @@
-﻿using Arcane.Operator.Models.Resources.Status.V1Alpha1;
+﻿using Arcane.Operator.Models.Base;
+using Arcane.Operator.Models.Resources.Status.V1Alpha1;
 using Arcane.Operator.Models.StreamDefinitions.Base;
 using Arcane.Operator.Services.Base;
+using Arcane.Operator.Services.Base.CommandHandlers;
 using Arcane.Operator.StreamingJobLifecycle;
 
 namespace Arcane.Operator.Models.Commands;
