+++ src/Models/Commands/StreamingJobCommands.cs
@@ -1,5 +1,5 @@
-﻿using Arcane.Operator.Models.StreamDefinitions.Base;
-using Arcane.Operator.Services.Base;
+﻿using Arcane.Operator.Models.Base;
+using Arcane.Operator.Models.StreamDefinitions.Base;
 using Arcane.Operator.StreamingJobLifecycle;
 using k8s.Models;
