+++ src/Services/Base/CommandHandlers/IStreamingJobCommandHandler.cs
@@ -1,7 +1,8 @@
-﻿using Arcane.Operator.Models.Commands;
+﻿using Arcane.Operator.Models.Base;
+using Arcane.Operator.Models.Commands;
 using k8s.Models;
 
-namespace Arcane.Operator.Services.Base;
+namespace Arcane.Operator.Services.Base.CommandHandlers;
 
 /// <summary>
 /// Command handler for streaming job commands
