+++ src/Services/Base/Repositories/CustomResources/IStreamClassRepository.cs
@@ -1,13 +1,11 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Akka.Util;
-using Arcane.Operator.Models;
 using Arcane.Operator.Models.Commands;
 using Arcane.Operator.Models.Resources.Status.V1Alpha1;
 using Arcane.Operator.Models.Resources.StreamClass.Base;
-using Arcane.Operator.Services.Base.Repositories.CustomResources;
 
-namespace Arcane.Operator.Services.Base;
+namespace Arcane.Operator.Services.Base.Repositories.CustomResources;
 
 /// <summary>
 /// The stream class storage abstraction
