+++ src/Services/Base/Repositories/CustomResources/IStreamingJobTemplateRepository.cs
