+++ src/Services/CommandHandlers/AnnotationCommandHandler.cs
@@ -2,8 +2,10 @@
 using Akka.Util;
 using Akka.Util.Extensions;
 using Arcane.Operator.Extensions;
+using Arcane.Operator.Models.Base;
 using Arcane.Operator.Models.StreamDefinitions.Base;
-using Arcane.Operator.Services.Base;
+using Arcane.Operator.Services.Base.CommandHandlers;
+using Arcane.Operator.Services.Base.Repositories.CustomResources;
 using k8s.Models;
 using Microsoft.Extensions.Logging;
 using Snd.Sdk.Kubernetes.Base;
