+++ src/Services/CommandHandlers/StreamingJobCommandHandler.cs
@@ -4,11 +4,14 @@ using System.Threading.Tasks;
 using Akka.Util;
 using Akka.Util.Extensions;
 using Arcane.Operator.Extensions;
+using Arcane.Operator.Models.Base;
 using Arcane.Operator.Models.Commands;
 using Arcane.Operator.Models.Resources.Status.V1Alpha1;
 using Arcane.Operator.Models.Resources.StreamClass.Base;
 using Arcane.Operator.Models.StreamDefinitions.Base;
 using Arcane.Operator.Services.Base;
+using Arcane.Operator.Services.Base.CommandHandlers;
+using Arcane.Operator.Services.Base.Repositories.CustomResources;
 using k8s.Models;
 using Microsoft.Extensions.Logging;
 using Snd.Sdk.Kubernetes;
