+++ src/Startup.cs
@@ -4,9 +4,11 @@ using System.Text.Json.Serialization;
 using Arcane.Operator.Configurations;
 using Arcane.Operator.Models.StreamDefinitions.Base;
 using Arcane.Operator.Services.Base;
+using Arcane.Operator.Services.Base.Metrics;
+using Arcane.Operator.Services.Base.Operators;
 using Arcane.Operator.Services.Base.Repositories.CustomResources;
 using Arcane.Operator.Services.Metrics;
-using Arcane.Operator.Services.Operator;
+using Arcane.Operator.Services.Operators;
 using Arcane.Operator.Services.Repositories.CustomResources;
 using Azure.Data.Tables;
 using Microsoft.AspNetCore.Builder;
