+++ test/Services/StreamOperatorServiceTests.cs
@@ -11,16 +11,20 @@ using Akka.Util.Extensions;
 using Arcane.Operator.Configurations;
 using Arcane.Operator.Configurations.Common;
 using Arcane.Operator.Models.Api;
+using Arcane.Operator.Models.Base;
 using Arcane.Operator.Models.Commands;
 using Arcane.Operator.Models.Resources.JobTemplates.Base;
 using Arcane.Operator.Models.Resources.StreamDefinitions;
 using Arcane.Operator.Models.StreamDefinitions.Base;
 using Arcane.Operator.Services.Base;
+using Arcane.Operator.Services.Base.CommandHandlers;
+using Arcane.Operator.Services.Base.Metrics;
+using Arcane.Operator.Services.Base.Operators;
 using Arcane.Operator.Services.Base.Repositories.CustomResources;
 using Arcane.Operator.Services.Base.Repositories.StreamingJob;
 using Arcane.Operator.Services.CommandHandlers;
 using Arcane.Operator.Services.Metrics;
-using Arcane.Operator.Services.Operator;
+using Arcane.Operator.Services.Operators;
 using Arcane.Operator.Services.Repositories.CustomResources;
 using Arcane.Operator.Services.Repositories.StreamingJob;
 using Arcane.Operator.StreamingJobLifecycle;
