+++ test/Services/TestCases/StreamDefinitionTestCases.cs
@@ -2,10 +2,8 @@
 using System.Collections.Generic;
 using System.Text.Json;
 using Arcane.Operator.Models.Resources.StreamDefinitions;
-using Arcane.Operator.Models.StreamDefinitions;
 using Arcane.Operator.Models.StreamDefinitions.Base;
 using Arcane.Operator.StreamingJobLifecycle;
-using Avro.IO.Parsing;
 using k8s.Models;
 
 namespace Arcane.Operator.Tests.Services.TestCases;
